career guidance/Backend/app/requirements.txt
fastapi==0.97.0
uvicorn==0.22.0
requests==2.31.0
beautifulsoup4==4.12.2
redis==4.5.5
fastapi-cache2==0.2.1
motor==3.2.0
apscheduler==3.10.1
httpx==0.24.1
python-dotenv==1.0.0